RZConfiguration - new instance var : logpath / path to where log files should go. With default values this is $RAZOR_HOME/log/
RZPersistController
- object_hash_update now returns the updated hash instead of whole array, helpful for RZData needs

RZPersistMongo
- object_doc_update now returns the updated hash instead of whole array, helpful for RZData needs

RZData
- new method: persist_object. Takes a RZObject-based Razor object and persists to the DB. Returns a copy of the object with persistence bindings in place to do object-based refresh and updates
- new method: delete_all_objects. Takes a collection symbol and drops all the docs like they are bad habits
- new method: delete_object. Takes an object (Must be RZObject-based) and remove it from collection.
- new method: delete_object_by_uuid. Takes a collection symbol and object uuid and removes it from collection. Handy for when you know the uuid but don't have the object itself.

Unit Tests
- Completed Node tests for RZData
- added and completed Model tests too
- and went ahead and finished Policy tests because I was feeling scrappy

+++ spec/base/rz_002_data_spec.rb
@@ -35,7 +35,13 @@ describe RZData do
 
     after(:all) do
       #Restore razor.conf back
-      FileUtils.mv("#{CONFIG_PATH}.backup", CONFIG_PATH, :force => true) if File.exists?("#{CONFIG_PATH}.backup")
+      if File.exists?("#{CONFIG_PATH}.backup")
+        File.delete(CONFIG_PATH)
+        FileUtils.mv("#{CONFIG_PATH}.backup", CONFIG_PATH, :force => true) if File.exists?("#{CONFIG_PATH}.backup")
+      else
+        write_config(default_config)
+      end
+
     end
 
     it "should load a config from config path(#{CONFIG_PATH}) on init" do
@@ -144,10 +150,10 @@ describe RZData do
       (1..NODE_COUNT).each do
       |x|
         temp_node = RZNode.new({:@name => "rspec_node_junk#{x}", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
-        temp_node._persist_ctrl = @data.persist_ctrl
+        temp_node = @data.persist_object(temp_node)
         @last_uuid = temp_node.uuid
         #(0..rand(10)).each do
-        @data.persist_ctrl.object_hash_update(temp_node.to_hash, :node)
+        temp_node.update_self
         #end
       end
     end
@@ -158,24 +164,53 @@ describe RZData do
     end
 
     it "should have a list of Nodes" do
-      nodes = @data.fetch_all(:node)
+      nodes = @data.fetch_all_objects(:node)
       nodes.count.should == NODE_COUNT
     end
 
     it "should get a single node by UUID" do
-      node = @data.fetch_by_uuid(:node, @last_uuid)
+      node = @data.fetch_object_by_uuid(:node, @last_uuid)
       node.is_a?(RZNode).should == true
 
-      node = @data.fetch_by_uuid(:node, "12345")
+      node = @data.fetch_object_by_uuid(:node, "12345")
       node.is_a?(NilClass).should == true
     end
 
-    it "should be able to add a new Node (does not exist)" do
+    it "should be able to add a new Node (does not exist) and update" do
+      temp_node = RZNode.new({:@name => "rspec_node_junk_new", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_node = @data.persist_object(temp_node)
+      temp_node.update_self
 
+      node = @data.fetch_object_by_uuid(:node, temp_node.uuid)
+      node.version.should == 2
+    end
+
+    it "should be able to delete a specific Node by uuid" do
+      temp_node = RZNode.new({:@name => "rspec_node_junk_delete_uuid", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_node = @data.persist_object(temp_node)
+      temp_node.update_self
+
+      node = @data.fetch_object_by_uuid(:node, temp_node.uuid)
+      node.version.should == 2
+
+      @data.delete_object_by_uuid(node._collection, node.uuid).should == true
+      @data.fetch_object_by_uuid(:node, node.uuid).should == nil
+    end
+
+    it "should be able to delete a specific Node by object" do
+      temp_node = RZNode.new({:@name => "rspec_node_junk_delete_object", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_node = @data.persist_object(temp_node)
+      temp_node.update_self
+
+      node = @data.fetch_object_by_uuid(:node, temp_node.uuid)
+      node.version.should == 2
+
+      @data.delete_object(node).should == true
+      @data.fetch_object_by_uuid(:node, node.uuid).should == nil
     end
 
     it "should be able to update Node attributes for existing Node" do
-      node = @data.fetch_by_uuid(:node, @last_uuid)
+      node = @data.fetch_object_by_uuid(:node, @last_uuid)
       node.is_a?(RZNode).should == true
       node.attributes_hash = {:hostname => "nick_weaver", :ip_address => "1.1.1.1", :iq => 160}
       node.update_self
@@ -183,7 +218,7 @@ describe RZData do
       node.attributes_hash["ip_address"].should == "1.1.1.1"
       node.attributes_hash["iq"].should == 160
 
-      node_confirm = @data.fetch_by_uuid(:node, @last_uuid)
+      node_confirm = @data.fetch_object_by_uuid(:node, @last_uuid)
       node_confirm.is_a?(RZNode).should == true
       node_confirm.attributes_hash["hostname"].should == "nick_weaver"
       node_confirm.attributes_hash["ip_address"].should == "1.1.1.1"
@@ -191,40 +226,245 @@ describe RZData do
     end
 
     it "should be able to update the LastState for existing Node" do
-      node = @data.fetch_by_uuid(:node, @last_uuid)
+      node = @data.fetch_object_by_uuid(:node, @last_uuid)
       node.is_a?(RZNode).should == true
       node.last_state = :nick
       node.update_self
       node.last_state.should == :nick
 
-      node_confirm = @data.fetch_by_uuid(:node, @last_uuid)
+      node_confirm = @data.fetch_object_by_uuid(:node, @last_uuid)
       node_confirm.is_a?(RZNode).should == true
       node_confirm.last_state = :nick
     end
 
     it "should be able to update the CurrentState for existing Node" do
-      node = @data.fetch_by_uuid(:node, @last_uuid)
+      node = @data.fetch_object_by_uuid(:node, @last_uuid)
       node.is_a?(RZNode).should == true
       node.current_state = :nick
       node.update_self
       node.current_state.should == :nick
 
-      node_confirm = @data.fetch_by_uuid(:node, @last_uuid)
+      node_confirm = @data.fetch_object_by_uuid(:node, @last_uuid)
       node_confirm.is_a?(RZNode).should == true
       node_confirm.current_state = :nick
     end
 
-
     it "should be able to update the NextState for existing Node" do
-      node = @data.fetch_by_uuid(:node, @last_uuid)
+      node = @data.fetch_object_by_uuid(:node, @last_uuid)
       node.is_a?(RZNode).should == true
       node.next_state = :nick
       node.update_self
       node.next_state.should == :nick
 
-      node_confirm = @data.fetch_by_uuid(:node, @last_uuid)
+      node_confirm = @data.fetch_object_by_uuid(:node, @last_uuid)
       node_confirm.is_a?(RZNode).should == true
       node_confirm.next_state = :nick
+    end
+
+    it "should be able to delete all Nodes" do
+      @data.delete_all_objects(:node)
+      @data.fetch_all_objects(:node).count.should == 0
+    end
+
+  end
+
+  describe ".Models" do
+
+    before(:all) do
+      @data = RZData.new
+
+      (1..NODE_COUNT).each do
+      |x|
+        temp_model = RZModel.new({:@name => "rspec_model_junk#{x}", :@model_type => :base, :@values_hash => {}})
+        temp_model = @data.persist_object(temp_model)
+        @last_uuid = temp_model.uuid
+        #(0..rand(10)).each do
+        temp_model.update_self
+        #end
+      end
+    end
+
+    after(:all) do
+      @data.persist_ctrl.object_hash_remove_all(:model).should == true
+      @data.teardown
+    end
+
+    it "should have a list of Models" do
+      models = @data.fetch_all_objects(:model)
+      models.count.should == NODE_COUNT
+    end
+
+    it "should get a single model by UUID" do
+      model = @data.fetch_object_by_uuid(:model, @last_uuid)
+      model.is_a?(RZModel).should == true
+
+      model = @data.fetch_object_by_uuid(:model, "12345")
+      model.is_a?(NilClass).should == true
+    end
+
+    it "should be able to add a new Model (does not exist) and update" do
+      temp_model = RZModel.new({:@name => "rspec_model_junk_new", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_model = @data.persist_object(temp_model)
+      temp_model.update_self
+
+      model = @data.fetch_object_by_uuid(:model, temp_model.uuid)
+      model.version.should == 2
+    end
+
+    it "should be able to delete a specific Model by uuid" do
+      temp_model = RZModel.new({:@name => "rspec_model_junk_delete_uuid", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_model = @data.persist_object(temp_model)
+      temp_model.update_self
+
+      model = @data.fetch_object_by_uuid(:model, temp_model.uuid)
+      model.version.should == 2
+
+      @data.delete_object_by_uuid(model._collection, model.uuid).should == true
+      @data.fetch_object_by_uuid(:model, model.uuid).should == nil
+    end
+
+    it "should be able to delete a specific Model by object" do
+      temp_model = RZModel.new({:@name => "rspec_model_junk_delete_object", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_model = @data.persist_object(temp_model)
+      temp_model.update_self
+
+      model = @data.fetch_object_by_uuid(:model, temp_model.uuid)
+      model.version.should == 2
+
+      @data.delete_object(model).should == true
+      @data.fetch_object_by_uuid(:model, model.uuid).should == nil
+    end
+
+    it "should be able to update Model attributes for existing Model" do
+      model = @data.fetch_object_by_uuid(:model, @last_uuid)
+      model.is_a?(RZModel).should == true
+      model.values_hash = {:hostname => "nick_weaver", :ip_address => "1.1.1.1", :iq => 160}
+      model.update_self
+      model.values_hash["hostname"].should == "nick_weaver"
+      model.values_hash["ip_address"].should == "1.1.1.1"
+      model.values_hash["iq"].should == 160
+
+      model_confirm = @data.fetch_object_by_uuid(:model, @last_uuid)
+      model_confirm.is_a?(RZModel).should == true
+      model_confirm.values_hash["hostname"].should == "nick_weaver"
+      model_confirm.values_hash["ip_address"].should == "1.1.1.1"
+      model_confirm.values_hash["iq"].should == 160
+    end
+
+    it "should be able to update the LastState for existing Model" do
+      model = @data.fetch_object_by_uuid(:model, @last_uuid)
+      model.is_a?(RZModel).should == true
+      model.model_type = :nick
+      model.update_self
+      model.model_type.should == :nick
+
+      model_confirm = @data.fetch_object_by_uuid(:model, @last_uuid)
+      model_confirm.is_a?(RZModel).should == true
+      model_confirm.model_type = :nick
+    end
+
+
+    it "should be able to delete all Models" do
+      @data.delete_all_objects(:model)
+      @data.fetch_all_objects(:model).count.should == 0
+    end
+
+  end
+
+  describe ".Policies" do
+
+    before(:all) do
+      @data = RZData.new
+
+      (1..NODE_COUNT).each do
+      |x|
+        temp_policy = RZPolicy.new({:@name => "rspec_policy_junk#{x}", :@policy_type => :base, :@model => :base})
+        temp_policy = @data.persist_object(temp_policy)
+        @last_uuid = temp_policy.uuid
+        #(0..rand(10)).each do
+        temp_policy.update_self
+        #end
+      end
+    end
+
+    after(:all) do
+      @data.persist_ctrl.object_hash_remove_all(:policy).should == true
+      @data.teardown
+    end
+
+    it "should have a list of Policies" do
+      policies = @data.fetch_all_objects(:policy)
+      policies.count.should == NODE_COUNT
+    end
+
+    it "should get a single policy by UUID" do
+      policy = @data.fetch_object_by_uuid(:policy, @last_uuid)
+      policy.is_a?(RZPolicy).should == true
+
+      policy = @data.fetch_object_by_uuid(:policy, "12345")
+      policy.is_a?(NilClass).should == true
+    end
+
+    it "should be able to add a new Policy (does not exist) and update" do
+      temp_policy = RZPolicy.new({:@name => "rspec_policy_junk_new", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_policy = @data.persist_object(temp_policy)
+      temp_policy.update_self
+
+      policy = @data.fetch_object_by_uuid(:policy, temp_policy.uuid)
+      policy.version.should == 2
+    end
+
+    it "should be able to delete a specific Policy by uuid" do
+      temp_policy = RZPolicy.new({:@name => "rspec_policy_junk_delete_uuid", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_policy = @data.persist_object(temp_policy)
+      temp_policy.update_self
+
+      policy = @data.fetch_object_by_uuid(:policy, temp_policy.uuid)
+      policy.version.should == 2
+
+      @data.delete_object_by_uuid(policy._collection, policy.uuid).should == true
+      @data.fetch_object_by_uuid(:policy, policy.uuid).should == nil
+    end
+
+    it "should be able to delete a specific Policy by object" do
+      temp_policy = RZPolicy.new({:@name => "rspec_policy_junk_delete_object", :@last_state => :idle, :@current_state => :idle, :@next_state => :policy_applied})
+      temp_policy = @data.persist_object(temp_policy)
+      temp_policy.update_self
+
+      policy = @data.fetch_object_by_uuid(:policy, temp_policy.uuid)
+      policy.version.should == 2
+
+      @data.delete_object(policy).should == true
+      @data.fetch_object_by_uuid(:policy, policy.uuid).should == nil
+    end
+
+    it "should be able to update Policy attributes for existing Policy" do
+      policy = @data.fetch_object_by_uuid(:policy, @last_uuid)
+      policy.is_a?(RZPolicy).should == true
+      policy.model = :nick
+      policy.update_self
+
+      policy_confirm = @data.fetch_object_by_uuid(:policy, @last_uuid)
+      policy_confirm.is_a?(RZPolicy).should == true
+      policy_confirm.model.should == :nick
+    end
+
+    it "should be able to update the LastState for existing Policy" do
+      policy = @data.fetch_object_by_uuid(:policy, @last_uuid)
+      policy.is_a?(RZPolicy).should == true
+      policy.policy_type = :nick
+      policy.update_self
+      policy.policy_type.should == :nick
+
+      policy_confirm = @data.fetch_object_by_uuid(:policy, @last_uuid)
+      policy_confirm.is_a?(RZPolicy).should == true
+      policy_confirm.policy_type = :nick
+    end
+
+
+    it "should be able to delete all Policies" do
+      @data.delete_all_objects(:policy)
+      @data.fetch_all_objects(:policy).count.should == 0
     end
 
   end
